Add the 24 hour logged path of the vehicle into Google map.
User can see a track of his vehicle for last 24 hours. Also he can see
the speed and other values for each location of the car for last 24
hours. ATTENTION!!! This is debug version. Current version do not
displays real track of the vehicle for last 24 hours because current
version of the OVMS server do not responses from C33 command correctly.
This version emulates the three points of the track.